extract abstract base driver classes for vendors
