+++ lib/Doctrine/DBAL/Driver/DrizzlePDOMySql/Driver.php
@@ -19,19 +19,15 @@
 
 namespace Doctrine\DBAL\Driver\DrizzlePDOMySql;
 
-use Doctrine\DBAL\Driver\DriverException;
-use Doctrine\DBAL\Driver\ExceptionConverterDriver;
-use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\Platforms\DrizzlePlatform;
 use Doctrine\DBAL\Schema\DrizzleSchemaManager;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
 
 /**
  * Drizzle driver using PDO MySql.
  *
  * @author Kim Hemsø Rasmussen <kimhemsoe@gmail.com>
  */
-class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, VersionAwarePlatformDriver
+class Driver extends \Doctrine\DBAL\Driver\PDOMySql\Driver
 {
     /**
      * {@inheritdoc}
@@ -39,39 +35,13 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     public function connect(array $params, $username = null, $password = null, array $driverOptions = array())
     {
         $conn = new Connection(
-            $this->_constructPdoDsn($params),
+            $this->constructPdoDsn($params),
             $username,
             $password,
             $driverOptions
         );
 
         return $conn;
-    }
-
-    /**
-     * Constructs the Drizzle MySql PDO DSN.
-     *
-     * @param array $params
-     *
-     * @return string The DSN.
-     */
-    private function _constructPdoDsn(array $params)
-    {
-        $dsn = 'mysql:';
-        if (isset($params['host']) && $params['host'] != '') {
-            $dsn .= 'host=' . $params['host'] . ';';
-        }
-        if (isset($params['port'])) {
-            $dsn .= 'port=' . $params['port'] . ';';
-        }
-        if (isset($params['dbname'])) {
-            $dsn .= 'dbname=' . $params['dbname'] . ';';
-        }
-        if (isset($params['unix_socket'])) {
-            $dsn .= 'unix_socket=' . $params['unix_socket'] . ';';
-        }
-
-        return $dsn;
     }
 
     /**
@@ -104,93 +74,5 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     public function getName()
     {
         return 'drizzle_pdo_mysql';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(\Doctrine\DBAL\Connection $conn)
-    {
-        $params = $conn->getParams();
-
-        return $params['dbname'];
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @link http://dev.mysql.com/doc/refman/5.7/en/error-messages-client.html
-     * @link http://dev.mysql.com/doc/refman/5.7/en/error-messages-server.html
-     */
-    public function convertException($message, DriverException $exception)
-    {
-        switch ($exception->getErrorCode()) {
-            case '1050':
-                return new Exception\TableExistsException($message, $exception);
-
-            case '1051':
-            case '1146':
-                return new Exception\TableNotFoundException($message, $exception);
-
-            case '1216':
-            case '1217':
-            case '1451':
-            case '1452':
-                return new Exception\ForeignKeyConstraintViolationException($message, $exception);
-
-            case '1062':
-            case '1557':
-            case '1569':
-            case '1586':
-                return new Exception\UniqueConstraintViolationException($message, $exception);
-
-            case '1054':
-            case '1166':
-            case '1611':
-                return new Exception\InvalidFieldNameException($message, $exception);
-
-            case '1052':
-            case '1060':
-            case '1110':
-                return new Exception\NonUniqueFieldNameException($message, $exception);
-
-            case '1064':
-            case '1149':
-            case '1287':
-            case '1341':
-            case '1342':
-            case '1343':
-            case '1344':
-            case '1382':
-            case '1479':
-            case '1541':
-            case '1554':
-            case '1626':
-                return new Exception\SyntaxErrorException($message, $exception);
-
-            case '1044':
-            case '1045':
-            case '1046':
-            case '1049':
-            case '1095':
-            case '1142':
-            case '1143':
-            case '1227':
-            case '1370':
-            case '2002':
-            case '2005':
-                return new Exception\ConnectionException($message, $exception);
-
-            case '1048':
-            case '1121':
-            case '1138':
-            case '1171':
-            case '1252':
-            case '1263':
-            case '1566':
-                return new Exception\NotNullConstraintViolationException($message, $exception);
-        }
-
-        return new Exception\DriverException($message, $exception);
     }
 }
