+++ lib/Doctrine/DBAL/Driver/OCI8/Driver.php
@@ -19,14 +19,8 @@
 
 namespace Doctrine\DBAL\Driver\OCI8;
 
-use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
-use Doctrine\DBAL\Driver\DriverException;
-use Doctrine\DBAL\Driver\ExceptionConverterDriver;
-use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Platforms\OraclePlatform;
-use Doctrine\DBAL\Schema\OracleSchemaManager;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
+use Doctrine\DBAL\Driver\AbstractOracleDriver;
 
 /**
  * A Doctrine DBAL driver for the Oracle OCI8 PHP extensions.
@@ -34,7 +28,7 @@ use Doctrine\DBAL\VersionAwarePlatformDriver;
  * @author Roman Borschel <roman@code-factory.org>
  * @since 2.0
  */
-class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, VersionAwarePlatformDriver
+class Driver extends AbstractOracleDriver
 {
     /**
      * {@inheritdoc}
@@ -64,64 +58,7 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
      */
     protected function _constructDsn(array $params)
     {
-        $dsn = '';
-        if (isset($params['host']) && $params['host'] != '') {
-            $dsn .= '(DESCRIPTION=(ADDRESS_LIST=(ADDRESS=(PROTOCOL=TCP)' .
-                   '(HOST=' . $params['host'] . ')';
-
-            if (isset($params['port'])) {
-                $dsn .= '(PORT=' . $params['port'] . ')';
-            } else {
-                $dsn .= '(PORT=1521)';
-            }
-
-            $serviceName = $params['dbname'];
-
-            if ( ! empty($params['servicename'])) {
-                $serviceName = $params['servicename'];
-            }
-
-            $service = 'SID=' . $serviceName;
-            $pooled   = '';
-
-            if (isset($params['service']) && $params['service'] == true) {
-                $service = 'SERVICE_NAME=' . $serviceName;
-            }
-
-            if (isset($params['pooled']) && $params['pooled'] == true) {
-                $pooled = '(SERVER=POOLED)';
-            }
-
-            $dsn .= '))(CONNECT_DATA=(' . $service . ')' . $pooled . '))';
-        } else {
-            $dsn .= $params['dbname'];
-        }
-
-        return $dsn;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function createDatabasePlatformForVersion($version)
-    {
-        return $this->getDatabasePlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabasePlatform()
-    {
-        return new OraclePlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSchemaManager(Connection $conn)
-    {
-        return new OracleSchemaManager($conn);
+        return $this->getEasyConnectString($params);
     }
 
     /**
@@ -130,56 +67,5 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     public function getName()
     {
         return 'oci8';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(Connection $conn)
-    {
-        $params = $conn->getParams();
-
-        return $params['user'];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function convertException($message, DriverException $exception)
-    {
-        switch ($exception->getErrorCode()) {
-            case '1':
-            case '2299':
-            case '38911':
-                return new Exception\UniqueConstraintViolationException($message, $exception);
-
-            case '904':
-                return new Exception\InvalidFieldNameException($message, $exception);
-
-            case '918':
-            case '960':
-                return new Exception\NonUniqueFieldNameException($message, $exception);
-
-            case '923':
-                return new Exception\SyntaxErrorException($message, $exception);
-
-            case '942':
-                return new Exception\TableNotFoundException($message, $exception);
-
-            case '955':
-                return new Exception\TableExistsException($message, $exception);
-
-            case '1017':
-            case '12545':
-                return new Exception\ConnectionException($message, $exception);
-
-            case '1400':
-                return new Exception\NotNullConstraintViolationException($message, $exception);
-
-            case '2292':
-                return new Exception\ForeignKeyConstraintViolationException($message, $exception);
-        }
-
-        return new Exception\DriverException($message, $exception);
     }
 }
