+++ lib/Doctrine/DBAL/Driver/PDOMySql/Driver.php
@@ -19,16 +19,9 @@
 
 namespace Doctrine\DBAL\Driver\PDOMySql;
 
-use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
-use Doctrine\DBAL\Driver\DriverException;
-use Doctrine\DBAL\Driver\ExceptionConverterDriver;
+use Doctrine\DBAL\Driver\AbstractMySQLDriver;
 use Doctrine\DBAL\Driver\PDOConnection;
-use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Platforms\MySQL57Platform;
-use Doctrine\DBAL\Platforms\MySqlPlatform;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
-use Doctrine\DBAL\Schema\MySqlSchemaManager;
 use PDOException;
 
 /**
@@ -36,7 +29,7 @@ use PDOException;
  *
  * @since 2.0
  */
-class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, VersionAwarePlatformDriver
+class Driver extends AbstractMySQLDriver
 {
     /**
      * {@inheritdoc}
@@ -45,7 +38,7 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     {
         try {
             $conn = new PDOConnection(
-                $this->_constructPdoDsn($params),
+                $this->constructPdoDsn($params),
                 $username,
                 $password,
                 $driverOptions
@@ -64,7 +57,7 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
      *
      * @return string The DSN.
      */
-    private function _constructPdoDsn(array $params)
+    protected function constructPdoDsn(array $params)
     {
         $dsn = 'mysql:';
         if (isset($params['host']) && $params['host'] != '') {
@@ -89,139 +82,8 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     /**
      * {@inheritdoc}
      */
-    public function createDatabasePlatformForVersion($version)
-    {
-        if ( ! preg_match('/^(?P<major>\d+)(?:\.(?P<minor>\d+)(?:\.(?P<patch>\d+))?)?/', $version, $versionParts)) {
-            throw DBALException::invalidPlatformVersionSpecified(
-                $version,
-                '<major_version>.<minor_version>.<patch_version>'
-            );
-        }
-
-        $majorVersion = $versionParts['major'];
-        $minorVersion = isset($versionParts['minor']) ? $versionParts['minor'] : 0;
-        $patchVersion = isset($versionParts['patch']) ? $versionParts['patch'] : 0;
-        $version      = $majorVersion . '.' . $minorVersion . '.' . $patchVersion;
-
-        if (version_compare($version, '5.7', '>=')) {
-            return new MySQL57Platform();
-        }
-
-        return $this->getDatabasePlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabasePlatform()
-    {
-        return new MySqlPlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSchemaManager(Connection $conn)
-    {
-        return new MySqlSchemaManager($conn);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getName()
     {
         return 'pdo_mysql';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(Connection $conn)
-    {
-        $params = $conn->getParams();
-
-        if (isset($params['dbname'])) {
-            return $params['dbname'];
-        }
-        return $conn->query('SELECT DATABASE()')->fetchColumn();
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @link http://dev.mysql.com/doc/refman/5.7/en/error-messages-client.html
-     * @link http://dev.mysql.com/doc/refman/5.7/en/error-messages-server.html
-     */
-    public function convertException($message, DriverException $exception)
-    {
-        switch ($exception->getErrorCode()) {
-            case '1050':
-                return new Exception\TableExistsException($message, $exception);
-
-            case '1051':
-            case '1146':
-                return new Exception\TableNotFoundException($message, $exception);
-
-            case '1216':
-            case '1217':
-            case '1451':
-            case '1452':
-                return new Exception\ForeignKeyConstraintViolationException($message, $exception);
-
-            case '1062':
-            case '1557':
-            case '1569':
-            case '1586':
-                return new Exception\UniqueConstraintViolationException($message, $exception);
-
-            case '1054':
-            case '1166':
-            case '1611':
-                return new Exception\InvalidFieldNameException($message, $exception);
-
-            case '1052':
-            case '1060':
-            case '1110':
-                return new Exception\NonUniqueFieldNameException($message, $exception);
-
-            case '1064':
-            case '1149':
-            case '1287':
-            case '1341':
-            case '1342':
-            case '1343':
-            case '1344':
-            case '1382':
-            case '1479':
-            case '1541':
-            case '1554':
-            case '1626':
-                return new Exception\SyntaxErrorException($message, $exception);
-
-            case '1044':
-            case '1045':
-            case '1046':
-            case '1049':
-            case '1095':
-            case '1142':
-            case '1143':
-            case '1227':
-            case '1370':
-            case '2002':
-            case '2005':
-                return new Exception\ConnectionException($message, $exception);
-
-            case '1048':
-            case '1121':
-            case '1138':
-            case '1171':
-            case '1252':
-            case '1263':
-            case '1566':
-                return new Exception\NotNullConstraintViolationException($message, $exception);
-        }
-
-        return new Exception\DriverException($message, $exception);
     }
 }
