+++ lib/Doctrine/DBAL/Driver/PDOOracle/Driver.php
@@ -19,15 +19,9 @@
 
 namespace Doctrine\DBAL\Driver\PDOOracle;
 
-use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
-use Doctrine\DBAL\Driver\DriverException;
-use Doctrine\DBAL\Driver\ExceptionConverterDriver;
+use Doctrine\DBAL\Driver\AbstractOracleDriver;
 use Doctrine\DBAL\Driver\PDOConnection;
-use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Platforms\OraclePlatform;
-use Doctrine\DBAL\Schema\OracleSchemaManager;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
 
 /**
  * PDO Oracle driver.
@@ -37,7 +31,7 @@ use Doctrine\DBAL\VersionAwarePlatformDriver;
  * which leads us to the recommendation to use the "oci8" driver to connect
  * to Oracle instead.
  */
-class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, VersionAwarePlatformDriver
+class Driver extends AbstractOracleDriver
 {
     /**
      * {@inheritdoc}
@@ -46,7 +40,7 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     {
         try {
             return new PDOConnection(
-                $this->_constructPdoDsn($params),
+                $this->constructPdoDsn($params),
                 $username,
                 $password,
                 $driverOptions
@@ -63,41 +57,9 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
      *
      * @return string The DSN.
      */
-    private function _constructPdoDsn(array $params)
+    private function constructPdoDsn(array $params)
     {
-        $dsn = 'oci:dbname=';
-
-        if (isset($params['host']) && $params['host'] != '') {
-            $dsn .= '(DESCRIPTION=(ADDRESS_LIST=(ADDRESS=(PROTOCOL=TCP)' .
-                '(HOST=' . $params['host'] . ')';
-
-            if (isset($params['port'])) {
-                $dsn .= '(PORT=' . $params['port'] . ')';
-            } else {
-                $dsn .= '(PORT=1521)';
-            }
-
-            $serviceName = $params['dbname'];
-
-            if ( ! empty($params['servicename'])) {
-                $serviceName = $params['servicename'];
-            }
-
-            $service = 'SID=' . $serviceName;
-            $pooled  = '';
-
-            if (isset($params['service']) && $params['service'] == true) {
-                $service = 'SERVICE_NAME=' . $serviceName;
-            }
-
-            if (isset($params['pooled']) && $params['pooled'] == true) {
-                $pooled = '(SERVER=POOLED)';
-            }
-
-            $dsn .= '))(CONNECT_DATA=(' . $service . ')' . $pooled . '))';
-        } else {
-            $dsn .= $params['dbname'];
-        }
+        $dsn = 'oci:dbname=' . $this->getEasyConnectString($params);
 
         if (isset($params['charset'])) {
             $dsn .= ';charset=' . $params['charset'];
@@ -109,83 +71,8 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     /**
      * {@inheritdoc}
      */
-    public function createDatabasePlatformForVersion($version)
-    {
-        return $this->getDatabasePlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabasePlatform()
-    {
-        return new OraclePlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSchemaManager(Connection $conn)
-    {
-        return new OracleSchemaManager($conn);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getName()
     {
         return 'pdo_oracle';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(Connection $conn)
-    {
-        $params = $conn->getParams();
-
-        return $params['user'];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function convertException($message, DriverException $exception)
-    {
-        switch ($exception->getErrorCode()) {
-            case '1':
-            case '2299':
-            case '38911':
-                return new Exception\UniqueConstraintViolationException($message, $exception);
-
-            case '904':
-                return new Exception\InvalidFieldNameException($message, $exception);
-
-            case '918':
-            case '960':
-                return new Exception\NonUniqueFieldNameException($message, $exception);
-
-            case '923':
-                return new Exception\SyntaxErrorException($message, $exception);
-
-            case '942':
-                return new Exception\TableNotFoundException($message, $exception);
-
-            case '955':
-                return new Exception\TableExistsException($message, $exception);
-
-            case '1017':
-            case '12545':
-                return new Exception\ConnectionException($message, $exception);
-
-            case '1400':
-                return new Exception\NotNullConstraintViolationException($message, $exception);
-
-            case '2292':
-                return new Exception\ForeignKeyConstraintViolationException($message, $exception);
-        }
-
-        return new Exception\DriverException($message, $exception);
     }
 }
