+++ lib/Doctrine/DBAL/Driver/PDOPgSql/Driver.php
@@ -19,24 +19,17 @@
 
 namespace Doctrine\DBAL\Driver\PDOPgSql;
 
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Driver\DriverException;
+use Doctrine\DBAL\Driver\AbstractPostgreSQLDriver;
 use Doctrine\DBAL\Driver\PDOConnection;
-use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Platforms\PostgreSQL92Platform;
-use Doctrine\DBAL\Platforms\PostgreSqlPlatform;
 use Doctrine\DBAL\DBALException;
-use Doctrine\DBAL\Schema\PostgreSqlSchemaManager;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
 use PDOException;
-use Doctrine\DBAL\Driver\ExceptionConverterDriver;
 
 /**
  * Driver that connects through pdo_pgsql.
  *
  * @since 2.0
  */
-class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, VersionAwarePlatformDriver
+class Driver extends AbstractPostgreSQLDriver
 {
     /**
      * {@inheritdoc}
@@ -50,7 +43,7 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
                 $password,
                 $driverOptions
             );
-        } catch(PDOException $e) {
+        } catch (PDOException $e) {
             throw DBALException::driverException($this, $e);
         }
     }
@@ -92,105 +85,8 @@ class Driver implements \Doctrine\DBAL\Driver, ExceptionConverterDriver, Version
     /**
      * {@inheritdoc}
      */
-    public function createDatabasePlatformForVersion($version)
-    {
-        if ( ! preg_match('/^(?P<major>\d+)(?:\.(?P<minor>\d+)(?:\.(?P<patch>\d+))?)?/', $version, $versionParts)) {
-            throw DBALException::invalidPlatformVersionSpecified(
-                $version,
-                '<major_version>.<minor_version>.<patch_version>'
-            );
-        }
-
-        $majorVersion = $versionParts['major'];
-        $minorVersion = isset($versionParts['minor']) ? $versionParts['minor'] : 0;
-        $patchVersion = isset($versionParts['patch']) ? $versionParts['patch'] : 0;
-        $version      = $majorVersion . '.' . $minorVersion . '.' . $patchVersion;
-
-        if (version_compare($version, '9.2', '>=')) {
-            return new PostgreSQL92Platform();
-        }
-
-        return $this->getDatabasePlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabasePlatform()
-    {
-        return new PostgreSqlPlatform();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSchemaManager(Connection $conn)
-    {
-        return new PostgreSqlSchemaManager($conn);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getName()
     {
         return 'pdo_pgsql';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(Connection $conn)
-    {
-        $params = $conn->getParams();
-
-        return (isset($params['dbname']))
-            ? $params['dbname']
-            : $conn->query('SELECT CURRENT_DATABASE()')->fetchColumn();
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @link http://www.postgresql.org/docs/9.3/static/errcodes-appendix.html
-     */
-    public function convertException($message, DriverException $exception)
-    {
-        switch ($exception->getSQLState()) {
-            case '23502':
-                return new Exception\NotNullConstraintViolationException($message, $exception);
-
-            case '23503':
-                return new Exception\ForeignKeyConstraintViolationException($message, $exception);
-
-            case '23505':
-                return new Exception\UniqueConstraintViolationException($message, $exception);
-
-            case '42601':
-                return new Exception\SyntaxErrorException($message, $exception);
-
-            case '42702':
-                return new Exception\NonUniqueFieldNameException($message, $exception);
-
-            case '42703':
-                return new Exception\InvalidFieldNameException($message, $exception);
-
-            case '42P01':
-                return new Exception\TableNotFoundException($message, $exception);
-
-            case '42P07':
-                return new Exception\TableExistsException($message, $exception);
-
-            case '7':
-                // In some case (mainly connection errors) the PDO exception does not provide a SQLSTATE via its code.
-                // The exception code is always set to 7 here.
-                // We have to match against the SQLSTATE in the error message in these cases.
-                if (strpos($exception->getMessage(), 'SQLSTATE[08006]') !== false) {
-                    return new Exception\ConnectionException($message, $exception);
-                }
-                break;
-        }
-
-        return new Exception\DriverException($message, $exception);
     }
 }
