+++ lib/Doctrine/DBAL/Driver/PDOSqlsrv/Driver.php
@@ -19,20 +19,14 @@
 
 namespace Doctrine\DBAL\Driver\PDOSqlsrv;
 
-use Doctrine\DBAL\DBALException;
-use Doctrine\DBAL\Platforms\SQLServer2005Platform;
-use Doctrine\DBAL\Platforms\SQLServer2008Platform;
-use Doctrine\DBAL\Platforms\SQLServer2012Platform;
-use Doctrine\DBAL\Platforms\SQLServerPlatform;
-use Doctrine\DBAL\Schema\SQLServerSchemaManager;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
+use Doctrine\DBAL\Driver\AbstractSQLServerDriver;
 
 /**
  * The PDO-based Sqlsrv driver.
  *
  * @since 2.0
  */
-class Driver implements \Doctrine\DBAL\Driver, VersionAwarePlatformDriver
+class Driver extends AbstractSQLServerDriver
 {
     /**
      * {@inheritdoc}
@@ -66,7 +60,7 @@ class Driver implements \Doctrine\DBAL\Driver, VersionAwarePlatformDriver
             $dsn .= ',' . $params['port'];
         }
 
-        if (isset($params['dbname'])) {;
+        if (isset($params['dbname'])) {
             $dsn .= ';Database=' .  $params['dbname'];
         }
 
@@ -80,76 +74,8 @@ class Driver implements \Doctrine\DBAL\Driver, VersionAwarePlatformDriver
     /**
      * {@inheritdoc}
      */
-    public function createDatabasePlatformForVersion($version)
-    {
-        if ( ! preg_match(
-            '/^(?P<major>\d+)(?:\.(?P<minor>\d+)(?:\.(?P<patch>\d+)(?:\.(?P<build>\d+))?)?)?/',
-            $version,
-            $versionParts
-        )) {
-            throw DBALException::invalidPlatformVersionSpecified(
-                $version,
-                '<major_version>.<minor_version>.<patch_version>.<build_version>'
-            );
-        }
-
-        $majorVersion = $versionParts['major'];
-        $minorVersion = isset($versionParts['minor']) ? $versionParts['minor'] : 0;
-        $patchVersion = isset($versionParts['patch']) ? $versionParts['patch'] : 0;
-        $buildVersion = isset($versionParts['build']) ? $versionParts['build'] : 0;
-        $version      = $majorVersion . '.' . $minorVersion . '.' . $patchVersion . $buildVersion;
-
-        switch(true) {
-            case version_compare($version, '11.00.2100', '>='):
-                return new SQLServer2012Platform();
-            case version_compare($version, '10.00.1600', '>='):
-                return new SQLServer2008Platform();
-            case version_compare($version, '9.00.1399', '>='):
-                return new SQLServer2005Platform();
-            default:
-                return new SQLServerPlatform();
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabasePlatform()
-    {
-        return new SQLServer2008Platform();
-    }
-    /**
-     * {@inheritdoc}
-     */
-
-    public function getSchemaManager(\Doctrine\DBAL\Connection $conn)
-    {
-        return new SQLServerSchemaManager($conn);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getName()
     {
         return 'pdo_sqlsrv';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDatabase(\Doctrine\DBAL\Connection $conn)
-    {
-        $params = $conn->getParams();
-
-        return $params['dbname'];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function convertExceptionCode(\Exception $exception)
-    {
-        return 0;
     }
 }
